requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-debug-toolbar==4.2.0
django-grappelli==3.0.8
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.24.0.post1
django-cors-headers==4.3.1
